Pulling function registration to a separate registry.
* registry interface is pulled out.
* Renaming some of the functions (transformX -> X)
* Function registry is no longer global during the execution of the uery (but it's not configurable yet)
