+++ query/function.go
@@ -17,180 +17,12 @@ package query
 import (
 	"fmt"
 	"math"
-	"strings"
 
 	"github.com/square/metrics/api"
 	"github.com/square/metrics/function"
-	"github.com/square/metrics/function/aggregate"
-	"github.com/square/metrics/function/filter"
-	"github.com/square/metrics/function/join"
 )
 
-var functionRegistry = map[string]MetricFunction{}
-
-func GetFunction(name string) (MetricFunction, bool) {
-	fun, ok := functionRegistry[name]
-	return fun, ok
-}
-
-func RegisterFunction(fun MetricFunction) error {
-	_, ok := functionRegistry[fun.Name]
-	if ok {
-		return fmt.Errorf("function %s has already been registered", fun.Name)
-	}
-	functionRegistry[fun.Name] = fun
-	return nil
-}
-
-func MustRegister(fun MetricFunction) {
-	err := RegisterFunction(fun)
-	if err != nil {
-		panic(fmt.Sprintf("function %s in failed to register", fun.Name))
-	}
-}
-
-type MetricFunction struct {
-	Name        string
-	MinArgument int
-	MaxArgument int
-	Groups      bool // Whether the function allows a 'group by' clause.
-	Compute     func(function.EvaluationContext, []function.Expression, []string) (function.Value, error)
-}
-
-func (f MetricFunction) Evaluate(
-	context function.EvaluationContext, arguments []function.Expression, groupBy []string,
-) (function.Value, error) {
-	// preprocessing
-	length := len(arguments)
-	if length < f.MinArgument || (f.MaxArgument != -1 && f.MaxArgument < length) {
-		return nil, ArgumentLengthError{f.Name, f.MinArgument, f.MaxArgument, length}
-	}
-	if len(groupBy) > 0 && !f.Groups {
-		return nil, fmt.Errorf("function %s doesn't allow a group-by clause", f.Name)
-	}
-	return f.Compute(context, arguments, groupBy)
-}
-
-func MakeOperatorMetricFunction(op string, operator func(float64, float64) float64) MetricFunction {
-	return MetricFunction{
-		Name:        op,
-		MinArgument: 2,
-		MaxArgument: 2,
-		Compute: func(context function.EvaluationContext, args []function.Expression, groups []string) (function.Value, error) {
-			leftValue, err := args[0].Evaluate(context)
-			if err != nil {
-				return nil, err
-			}
-			rightValue, err := args[1].Evaluate(context)
-			if err != nil {
-				return nil, err
-			}
-			leftList, err := leftValue.ToSeriesList(context.Timerange)
-			if err != nil {
-				return nil, err
-			}
-			rightList, err := rightValue.ToSeriesList(context.Timerange)
-			if err != nil {
-				return nil, err
-			}
-
-			joined := join.Join([]api.SeriesList{leftList, rightList})
-
-			result := make([]api.Timeseries, len(joined.Rows))
-
-			for i, row := range joined.Rows {
-				left := row.Row[0]
-				right := row.Row[1]
-				array := make([]float64, len(left.Values))
-				for j := 0; j < len(left.Values); j++ {
-					array[j] = operator(left.Values[j], right.Values[j])
-				}
-				result[i] = api.Timeseries{array, row.TagSet}
-			}
-
-			return function.SeriesListValue(api.SeriesList{
-				Series:    result,
-				Timerange: context.Timerange,
-				Name:      fmt.Sprintf("(%s %s %s)", leftValue.GetName(), op, rightValue.GetName()),
-			}), nil
-		},
-	}
-}
-
-// MakeAggregateMetricFunction takes a named aggregating function `[float64] => float64` and makes it into a MetricFunction.
-func MakeAggregateMetricFunction(name string, aggregator func([]float64) float64) MetricFunction {
-	return MetricFunction{
-		Name:        name,
-		MinArgument: 1,
-		MaxArgument: 1,
-		Groups:      true,
-		Compute: func(context function.EvaluationContext, args []function.Expression, groups []string) (function.Value, error) {
-			argument := args[0]
-			value, err := argument.Evaluate(context)
-			if err != nil {
-				return nil, err
-			}
-			seriesList, err := value.ToSeriesList(context.Timerange)
-			if err != nil {
-				return nil, err
-			}
-			result := aggregate.AggregateBy(seriesList, aggregator, groups)
-			groupNames := make([]string, len(groups))
-			for i, group := range groups {
-				groupNames[i] += group
-			}
-			if len(groups) == 0 {
-				result.Name = fmt.Sprintf("%s(%s)", name, value.GetName())
-			} else {
-				result.Name = fmt.Sprintf("%s(%s group by %s)", name, value.GetName(), strings.Join(groupNames, ", "))
-			}
-			return function.SeriesListValue(result), nil
-		},
-	}
-}
-
-// MakeTransformMetircFunction takes a named transforming function `[float64], [value] => [float64]` and makes it into a MetricFunction.
-func MakeTransformMetricFunction(name string, parameterCount int, transformer func([]float64, []function.Value, float64) ([]float64, error)) MetricFunction {
-	return MetricFunction{
-		Name:        name,
-		MinArgument: parameterCount + 1,
-		MaxArgument: parameterCount + 1,
-		Compute: func(context function.EvaluationContext, args []function.Expression, groups []string) (function.Value, error) {
-			// ApplyTransform(list api.SeriesList, transform transform, parameters []value) (api.SeriesList, error)
-			listValue, err := args[0].Evaluate(context)
-			if err != nil {
-				return nil, err
-			}
-			list, err := listValue.ToSeriesList(context.Timerange)
-			if err != nil {
-				return nil, err
-			}
-			parameters := make([]function.Value, parameterCount)
-			for i := range parameters {
-				parameters[i], err = args[i+1].Evaluate(context)
-				if err != nil {
-					return nil, err
-				}
-			}
-			result, err := ApplyTransform(list, transformer, parameters)
-			if err != nil {
-				return nil, err
-			}
-			parameterNames := make([]string, len(parameters))
-			for i, param := range parameters {
-				parameterNames[i] = param.GetName()
-			}
-			if len(parameters) != 0 {
-				result.Name = fmt.Sprintf("%s(%s, %s)", name, listValue.GetName(), strings.Join(parameterNames, ", "))
-			} else {
-				result.Name = fmt.Sprintf("%s(%s)", name, listValue.GetName())
-			}
-			return function.SeriesListValue(result), nil
-		},
-	}
-}
-
-var TimeshiftFunction = MetricFunction{
+var TimeshiftFunction = function.MetricFunction{
 	Name:        "transform.timeshift",
 	MinArgument: 2,
 	MaxArgument: 2,
@@ -220,7 +52,7 @@ var TimeshiftFunction = MetricFunction{
 	},
 }
 
-var MovingAverageFunction = MetricFunction{
+var MovingAverageFunction = function.MetricFunction{
 	Name:        "transform.moving_average",
 	MinArgument: 2,
 	MaxArgument: 2,
@@ -296,7 +128,7 @@ var MovingAverageFunction = MetricFunction{
 	},
 }
 
-var AliasFunction = MetricFunction{
+var AliasFunction = function.MetricFunction{
 	Name:        "transform.alias",
 	MinArgument: 2,
 	MaxArgument: 2,
@@ -320,39 +152,4 @@ var AliasFunction = MetricFunction{
 		list.Name = name
 		return function.SeriesListValue(list), nil
 	},
-}
-
-func MakeFilterMetricFunction(name string, summary func([]float64) float64, ascending bool) MetricFunction {
-	return MetricFunction{
-		Name:        name,
-		MinArgument: 2,
-		MaxArgument: 2,
-		Compute: func(context function.EvaluationContext, arguments []function.Expression, groups []string) (function.Value, error) {
-			value, err := arguments[0].Evaluate(context)
-			if err != nil {
-				return nil, err
-			}
-			// The value must be a SeriesList.
-			list, err := value.ToSeriesList(context.Timerange)
-			if err != nil {
-				return nil, err
-			}
-			countValue, err := arguments[1].Evaluate(context)
-			if err != nil {
-				return nil, err
-			}
-			countFloat, err := countValue.ToScalar()
-			if err != nil {
-				return nil, err
-			}
-			// Round to the nearest integer.
-			count := int(countFloat + 0.5)
-			if count < 0 {
-				return nil, fmt.Errorf("expected positive count but got %d", count)
-			}
-			result := filter.FilterBy(list, count, summary, ascending)
-			result.Name = fmt.Sprintf("%s(%s, %d)", name, value.GetName(), count)
-			return function.SeriesListValue(result), nil
-		},
-	}
 }
